+++ packages/pdf_graphics/lib/pdf_graphics.dart
@@ -5,6 +5,8 @@ library;
 
 export 'src/color.dart';
 export 'src/color_space.dart';
+export 'src/colorants.dart';
+export 'src/overprint_compositor.dart';
 export 'src/calibrated_color.dart';
 export 'src/document_ai.dart';
 export 'package:pdf_cos/pdf_cos.dart'